fix: trivial updates to the LS Parsers
- deprecate the LsRsyslogErrorfile which is no longer used by rules
- move FilePermissions to insights.parsers.ls from insights.util
- deprecate config_file_perms module and introduce LsFilePermissions
