+++ insights/specs/default.py
@@ -507,6 +507,7 @@ class DefaultSpecs(Specs):
     ls_lanRL = command_with_args('/bin/ls -lanRL %s', ls.list_with_lanRL, keep_rc=True)
     ls_laRZ = command_with_args('/bin/ls -laRZ %s', ls.list_with_laRZ, keep_rc=True)
     ls_laZ = command_with_args('/bin/ls -laZ %s', ls.list_with_laZ, keep_rc=True)
+    ls_file_perms = command_with_args('/bin/ls -lH %s', ls.list_file_perms, keep_rc=True)
     # Useful individual `ls` Specs
     ls_boot = simple_command("/bin/ls -lanR /boot")
     ls_dev = simple_command("/bin/ls -lanR /dev")
