+++ insights/tests/parsers/test_ls_file_listing.py
@@ -2,7 +2,7 @@
 import doctest
 
 from insights.parsers import ls as ls_module
-from insights.parsers.ls import FileListing
+from insights.parsers.ls import FileListingParser, LSlan, LsFilePermissions
 from insights.tests import context_wrap
 
 SINGLE_DIRECTORY = """
@@ -131,6 +131,10 @@ FILE_LISTING_DOC = '''
         lrwxrwxrwx.  1 0 0   17 Jul  6 23:32 grub -> /etc/default/grub
 '''
 
+LS_FILE_PERMISSIONS_DOC = '''
+-rw-r--r--. 1 root  root      46 Apr 24  2024 /etc/redhat-release
+-rw-r--r--. 1 liuxc wheel 664118 Feb 20 14:40 /var/log/messages
+'''
 # Note - should we test for anomalous but parseable entries?  E.g. block
 # devices without a major/minor number?  Or non-devices that have a comma in
 # the size?  Permissions that don't make sense?  Dates that don't make sense
@@ -140,7 +144,7 @@ FILE_LISTING_DOC = '''
 def test_single_directory():
     # ctx = context_wrap(SINGLE_DIRECTORY, path='ls_-la_.etc.pki.tls')
     ctx = context_wrap(SINGLE_DIRECTORY)
-    dirs = FileListing(ctx)
+    dirs = FileListingParser(ctx)
 
     assert '/etc/pki/tls' in dirs
     assert '/etc/pki/tls/certs' not in dirs
@@ -157,7 +161,7 @@ def test_single_directory():
 
 
 def test_multiple_directories():
-    dirs = FileListing(context_wrap(MULTIPLE_DIRECTORIES, path='ls_-la_.etc'))
+    dirs = FileListingParser(context_wrap(MULTIPLE_DIRECTORIES, path='ls_-la_.etc'))
 
     assert '/etc/sysconfig' in dirs
     assert '/etc/rc.d/rc3.d' in dirs
@@ -339,7 +343,7 @@ def test_raw_entry_of_permissions_of():
 
 
 def test_complicated_directory():
-    dirs = FileListing(context_wrap(COMPLICATED_FILES))
+    dirs = FileListingParser(context_wrap(COMPLICATED_FILES))
 
     # Test the things we expect to be different:
     listing = dirs.listing_of('/tmp')
@@ -394,7 +398,7 @@ def test_complicated_directory():
 
 
 def test_selinux_directory():
-    dirs = FileListing(context_wrap(SELINUX_DIRECTORY))
+    dirs = FileListingParser(context_wrap(SELINUX_DIRECTORY))
 
     # Test that one entry is exactly what we expect it to be.
     expected = {
@@ -415,7 +419,7 @@ def test_selinux_directory():
 
 
 def test_files_created_with_selinux_disabled():
-    dirs = FileListing(context_wrap(FILES_CREATED_WITH_SELINUX_DISABLED))
+    dirs = FileListingParser(context_wrap(FILES_CREATED_WITH_SELINUX_DISABLED))
 
     # Test that one entry is exactly what we expect it to be.
     assert dirs.dir_entry('/dev/mapper', 'lv_cpwtk001_data01') == {
@@ -434,6 +438,9 @@ def test_files_created_with_selinux_disabled():
 
 
 def test_doc_example():
-    env = {'ls_lan': FileListing(context_wrap(FILE_LISTING_DOC))}
+    env = {
+        'ls_lan': LSlan(context_wrap(FILE_LISTING_DOC)),
+        'ls_perms': LsFilePermissions(context_wrap(LS_FILE_PERMISSIONS_DOC)),
+    }
     failed, total = doctest.testmod(ls_module, globs=env)
     assert failed == 0
